refactor: create generic trap interface library for x86_64 backend support
As part of the x86_64 backend development (PR #476), this commit refactors trap handling
to be architecture-independent:

- Create dedicated 'trap' library at libs/trap/ for architecture-independent trap types
- Move Trap, Interrupt, and Exception enums from RISC-V-specific code to shared library
- Update RISC-V library to import trap types from the new shared library
- Update kernel trap_handler to use the generic trap interface
- Add trap library to workspace dependencies

This refactoring enables code sharing between RISC-V and future x86_64 implementations,
eliminating duplication and creating a clean architecture-agnostic interface for trap handling.

Verified: RISC-V kernel boots and runs successfully with these changes.

Related to: https://github.com/JonasKruckenberg/k23/pull/476
