+++ kernel/Cargo.toml
@@ -12,6 +12,7 @@ harness = false
 [dependencies]
 loader-api.workspace = true
 cpu-local.workspace = true
+trap.workspace = true
 spin = { workspace = true, features = ["thread-local", "lock_api"] }
 linked-list.workspace = true
 unwind2.workspace = true
