+++ kernel/src/mem/trap_handler.rs
@@ -9,7 +9,7 @@ use crate::mem::VirtualAddress;
 use crate::state::global;
 use core::ops::ControlFlow;
 use kasync::scheduler::Schedule;
-use riscv::scause::Trap;
+use trap::Trap;
 
 pub fn handle_page_fault(_trap: Trap, _tval: VirtualAddress) -> ControlFlow<()> {
     let current_task = global().executor.cpu_local_scheduler().current_task();
