+++ libs/trap/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "trap"
+description = "Generic trap/exception handling types"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# No external dependencies needed for basic trap types
